pyproject.toml
[project]
name = "skillet"
version = "0.1.0"
description = "Evaluation-driven Claude Code skill development"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cyclopts>=3.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "claude-agent-sdk",
]

[project.scripts]
skillet = "skillet.cli.main:main"

[project.optional-dependencies]
dev = [
    "rust-just>=1.0",
    "ruff>=0.8",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.24",
    "pytest-describe>=2.0",
    "pytest-watcher>=0.4",
    "pre-commit>=4.0",
    "bandit>=1.7",
    "ty>=0.0.1a7",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["skillet"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["skillet"]
python_files = ["*_test.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["skillet"]
branch = true
omit = ["*_test.py"]

[tool.coverage.report]
fail_under = 3
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

